requirements.txt
flask
flask-jwt-extended
python-dotenv
psycopg2
bcrypt
